Dependency fix.
If a user would like to create the sphinx documentation for the package,
they use the sphinx command but this is really optional. The setup file
has been modified to reflect this optionality.

The earlier version would cause an installation failure unless sphinx
was installed, which is not needed for normal installations.