Fix for redis-py's laxness.
The tests started failing with newer versions of redis-py. It used to be
that redis-py would choke on trying to access /dev/null but the newer
releases decide that because /dev/null does not begin with unix: then it
must be a redis: value and try to access the host
accordingly. Ultimately, redis-py connects to localhost with the default
redis port.

See https://github.com/andymccurdy/redis-py/issues/728.